sw: fix compiler warnings in unchanged apps
A clean build of every app showed warnings that the code changes in
this pass did not introduce. Fix them without changing behavior:

- restore_window_stress, wfi_drain_mepc_test: cast the 32-bit DDR
  address arithmetic through uintptr_t before making a pointer
  (-Wint-to-pointer-cast).
- ras_stress_test: drop expected_long, a reference value nothing read.
- plic_test: drop its MIP_MEIP, which redefined csr.h's (both 0x800).
- linux_irq_active_ddr_test, pde_return_hazard: mark the parameters of
  the naked assembly helpers unused, as mret_drain_deadlock does.
- coremark_pro: declare the port's exit() noreturn, which its comment
  already claimed, so __stack_chk_fail no longer "returns".

All of these apps build to byte-identical images before and after,
now without warnings. Two warnings remain in the sweep:
mret_timer_resume_test's -Wpointer-to-int-cast, from its 32-bit mepc
comparison, and sw/lib/src/memory.c's -Wtype-limits in the C library.

+++ sw/apps/restore_window_stress/main.c
@@ -313,7 +313,8 @@ int main(void)
          * evicted: its write-back drains below the L1D and the window's PT_EPC
          * load and SC miss cold. That refill in turn evicts the dirty alias
          * line, keeping a write-back draining inside the window. */
-        volatile rw_word_t *frame = (volatile rw_word_t *) (FRAME_BASE + ((i & 63u) << 6));
+        volatile rw_word_t *frame =
+            (volatile rw_word_t *) (uintptr_t) (FRAME_BASE + ((i & 63u) << 6));
         volatile rw_word_t *alias = (volatile rw_word_t *) ((uintptr_t) frame ^ FRAME_ALIAS_XOR);
 
         uint32_t to_umode = i & 1u;
